Rename vendor to third_party
This commit renames vendor to third_party because
the usage of the vendor directory in this repo is not
consistent with Go's conventions for the vendor directory
(https://golang.org/src/README.vendor).  Specifically,
the vendor directory contains two third-party packages
written in C, scamper and libparistraceroute, where the
first one is included as a source tarball and the second
one as a Git submodule.

The changes were successfully tested by running:

docker build .