Clean up Go source code
This commit makes changes to this repo's Go source files (excluding
the "*_test.go" files) so they would pass the following tools that
check for correctness and style:

  gofmt
  goimports
  golint
  go vet
  staticcheck

There is only one remaining "golint" warning in "parser/pt.go"
but we can ignore it because this file will soon be deprecated.

There is no code change in this commit that would result in a
different behavior.