Remove the remaining references to the pipe package
This commit removes the remaining references to the pipe package
and simplifies the code that uses the shx package.

This commit also adds some flags to make it easier to configure
scamper without having to edit code and rebuild traceroute-caller
for different configurations.  Specifically, "-O ptr" can be
enabled/disabled and tracelb's "-W" flag can be specified.

Run "./traceroute-caller -h" and/or see docker-compose.yaml for
an example of running with values different than the defaults.

The changes were tested locally with "go test ./..." and
docker-compose.