Use the shx package instead of the pipe package (#107)
* Use the shx package instead of the pipe package

The pipe package (gopkg.in/m-lab/pipe.v3) doesn't work as expected
because not only it doesn't support contexts its RunTimeout()
method does not correctly timeout.  As a result, the latency
numbers cannot be trusted.

This commit replaces the pipe package with the shx package in the
ScamperDaemon.trace().	Once we have validated the new code works
as expected, the Scamper.trace() will also use the shx package.

Because there are multiple traces being performed in parallel,
the new code improves log messages by including the context
identifier so individual contexts can be easily identified.

Tested the changes locally using docker-compose.

* Change the exepcted error message in scamper_test.go

* Remove unnecessary variable declaration

* Remove the remaining references to the pipe package

This commit removes the remaining references to the pipe package
and simplifies the code that uses the shx package.

This commit also adds some flags to make it easier to configure
scamper without having to edit code and rebuild traceroute-caller
for different configurations.  Specifically, "-O ptr" can be
enabled/disabled and tracelb's "-W" flag can be specified.

Run "./traceroute-caller -h" and/or see docker-compose.yaml for
an example of running with values different than the defaults.

The changes were tested locally with "go test ./..." and
docker-compose.

* Make changes suggested in code review