Remove code related to scamper daemon mode
The scamper tool used by traceroute-caller to run traceroutes can
operate in two modes: stand-alone binary mode and daemon mode.
In the stand-alone mode, an instance of scamper is invoked for
each traceroute run.  In the daemon mode, scamper is invoked once
as a daemon to listen on a socket for requests.  The daemon mode,
however, does not work properly and times out when there are many
concurrent requests.  As a result, about 75% of all traceroutes
were timing out.  We are now running scamper in stand-alone mode
and are not supporting the daemon mode anymore.

The changes were tested locally using docker-compose.