Refactor traceroute-caller source code
This commit refactors traceroute-caller code to make the code
easier to read and maintain, and also to make it more conformant
with Go style guidelines.

There is no change in the application logic.

This is a work in progress and will be followed by other commits.
Specifically, package testing code that is missing from this
commit will be added soon.

The main purpose of this commit is to facilitate the review
process.