Add NewScamper() to create a new scamper instance
This commit adds a new function, NewScamper(), to the tracer
package for creating a new instance of scamper based on the
configuration specified in the call if the configuration is valid.
The new scamper instance contains all of the information required
for invoking the scamper binary to run a traceroute.

The changes were tested locally with go test and docker-compose.