Add bigquery tags to struct fields and create an internal directory
This commit adds the missing bigquery tags to all struct fields of
datatypes that are stored in BigQuery.  In the legacy pipeline, we
used streaming inserts, and BigQuery respected the bigquery tags.
In Gardener 2.0, the rows are written to json files, and the
json tags are used.  Gardener then triggers the loading into
BigQuery from the json files.

This commit also moves ipcache and triggertrace packages to the
internal directory because these are not used externally.

Changes were tested locally with "go test -race" and "docker-compose".