Remove the notion of TCP socket cookies
The Linux kernel provides a unique cookie for each TCP socket
connection.  The cookie value is used by the tcp-info package to
generate a global UUID that is passed to tcp-info's eventsocket
clients.  Therefore, there is no reason for clients to know
about cookies or generate UUIDs from them.

This commit removes the notion of TCP socket cookies from
traceroute-caller.

Tested the changes locally with docker-compose.