Add load balancer IPs to "localIPs" (#167)
* Adds a function to add load balancer IPs to localIPs variable

* Adds log statement when a LB IP is added to localIPs

* Causes loadbalancerIPs() to work for both IP and CIDR formats

The values being parsed were ultimately fetched from the GCE VM metadata
server. The server seems to return IPv4 addresses in standard dotted decimal
notation, while IPv6 addresses are returned in CIDR format.

* Log errors from findDestination()

findDestination() returns an err, but it was previously unused other than to
check whether it was nil or not. This change just prints the error to help in
debugging.

* Bumps Go version to v1.21

* Adds loadbalancer IPs to handler.LocalIPs

This allows traceroute-caller to recognize the loadbalancer IPs as "local".
Without this traceroute-caller will fail to recognize its own (loadbalanced) IP
addresses and will not run any traces.

* Removes pointer from slice-of-pointers h.LocalIPs of type []*net.IP

Instead it is now just []net.IP, which is easier to work with.

* Updates Go to v1.21 in Travis-CI config

* Reverts Travis-CI config Go version back to v1.20 (1.21 not supported?)

* Sets Go version in Travis config back to 1.21, but in double quotes

* Reverts back to Go v1.20 in Travis-CI config (1.21 not supported)